Make private cache attributes not enumerable
When not eager-loading relationships, these are defined as properties
with a getter that only builds the relationship data structure if
requested. Once built, it keeps the built data structure in what can be
called a private attribute of sorts.

This commit make these private attributes not enumerable. This has a
number of benefits, mainly around not having this attributes show up
when listing all keys of the object, which in turn helps with
comparisons of this object with others, etc.